Refactor: move build-only NanVix staging into nanvix_build_common (review)
Addresses @ppenna's review on PR #519: build-only logic lived in
nanvix_common, which is a runtime dependency of nanvix_runner, so it (and a
panic) was compiled into the shipping binary.

- New build-only crate nanvix_build_common (never linked into runtime). Moves
  ArtifactKind, artifact_rel_paths, copy_artifacts_to_target,
  emit_rerun_for_copied_artifacts, and the snapshot tests out of nanvix_common.
- copy_artifacts_to_target / remove_stale_snapshot now return io::Result instead
  of panicking; the build-path caller (stage_artifacts_next_to_exe) decides to
  fail. nanvix_common stays panic-free.
- resolve_bin_dir centralizes NANVIX_BIN resolution (returns (dir, prefetched)
  tuple; treats empty NANVIX_BIN as unset; uses std::path::absolute).
- stage_artifacts_next_to_exe dedupes the identical copy/trust/rerun logic that
  was duplicated in wxc/build.rs and lxc/build.rs.
- nanvix_common keeps only runtime constants + types.
- Update copilot-instructions with the new crate.
