+++ src/core/lxc/Cargo.toml
@@ -13,7 +13,7 @@ microvm = ["dep:nanvix_binaries", "nanvix_binaries/microvm", "dep:nanvix_runner"
 
 [build-dependencies]
 mxc_build_common.workspace = true
-nanvix_common = { path = "../../backends/nanvix/common" }
+nanvix_build_common = { path = "../../backends/nanvix/build_common" }
 
 [dependencies]
 wxc_common = { workspace = true }
